Fixed minor leaks and fixed problem with multiple GUIs
I'm currently also trying to solve a bug where the memory increases by 15mb every time you exit fullscreen mode. No memory leaks are detected by the APIs or the standard library which is weird. I believe this is a bug with the DXGI API because I haven't been able to find any source of the leak other than the method IDXGISwapChain1::SetFullscreenState.